ml: training loop for singular internal node finished
A looot of regularization and hyperparams tweaking resulting in training
that has 70-80% on val data depending on number of classes. Works with
multiple hierarchy nodes. For validation, we do a lot of random data
augmentation on the fly, also interpolating between labels and samples
with dynamic alpha based on norm of gradients and gap betwen accuracy on
train and val.

All the tuning runs are visible on wandb.